% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{bib_splited_to_tibble}
\alias{bib_splited_to_tibble}
\title{Normalize column names across formats}
\usage{
bib_splited_to_tibble(bib_splited_by_field)
}
\description{
Normalize column names across formats
}
\keyword{internal}
