% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryPropMatrix.R
\name{binaryProbMatrix}
\alias{binaryProbMatrix}
\title{Binary Probability Matrix}
\usage{
binaryProbMatrix(nVar = 3, asList = TRUE)
}
\arguments{
\item{nVar}{number of variables, Default: 3}

\item{asList}{return as a list of matrix or one matrix, Default: TRUE}
}
\value{
list of matrix or matrix
}
\description{
Create a matrix with n columns filled with zeros and ones, as the probabilities of occurrences.
}
\examples{
if(interactive()){
 binaryPropMatrix(nvar=4)
 binaryPropMatrix(nvar=4, asList=F)
 }
}
