% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_trajectory_coherence.R
\name{.cosine_distance}
\alias{.cosine_distance}
\title{1 - Salton cosine between every pair of node profile vectors}
\usage{
.cosine_distance(m)
}
\description{
1 - Salton cosine between every pair of node profile vectors
}
\keyword{internal}
