% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_trajectory_coherence.R
\name{.feature_count_matrix}
\alias{.feature_count_matrix}
\title{Sparse node x feature count matrix from a (node, document_id, feature) frame}
\usage{
.feature_count_matrix(nf)
}
\description{
Cell (n, f) = number of documents in node n carrying feature f.
}
\keyword{internal}
