% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_trajectory_formation.R
\name{.formation_from_flow}
\alias{.formation_from_flow}
\title{Formation of a trajectory from the flow tree (its direct tributaries)}
\usage{
.formation_from_flow(flow, target, min_papers = 5, min_prop = 0.05)
}
\description{
Formation of a trajectory from the flow tree (its direct tributaries)
}
\keyword{internal}
