% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_influence.R
\name{.influence_core}
\alias{.influence_core}
\title{Indices, balance and net flow from a cross-citation matrix}
\usage{
.influence_core(C, glev)
}
\description{
Indices, balance and net flow from a cross-citation matrix
}
\keyword{internal}
