% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_influence.R
\name{.influence_matrix}
\alias{.influence_matrix}
\title{Cross-citation matrix from citing/cited group vectors}
\usage{
.influence_matrix(gi, gj, glev)
}
\description{
Cross-citation matrix from citing/cited group vectors
}
\keyword{internal}
