% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_influence.R
\name{.influence_null}
\alias{.influence_null}
\title{Per-channel permutation p-value matrix}
\usage{
.influence_null(C, gmap, edges, self, glev, null_reps)
}
\description{
Per-channel permutation p-value matrix
}
\keyword{internal}
