% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories_core.R
\name{.na_to_zero}
\alias{.na_to_zero}
\title{Replace NA values with zero}
\usage{
.na_to_zero(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector with NA values replaced by 0
}
\description{
Replace NA values with zero
}
\keyword{internal}
