% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_trajectory_dynamics.R
\name{.robust_z}
\alias{.robust_z}
\title{Outlier-robust z-score (modified z): (x - median) / MAD, NA -> 0}
\usage{
.robust_z(x)
}
\description{
Outlier-robust z-score (modified z): (x - median) / MAD, NA -> 0
}
\keyword{internal}
