% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components}
\alias{get_components}
\title{get_components}
\usage{
get_components(net, name_pad_width = 2)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{name_pad_width}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{component_distribution}},\code{\link[igraph]{V}},\code{\link[igraph]{as_data_frame}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{context}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{rename}}
\code{\link[stringr]{str_pad}}
\code{\link[tidygraph]{as_tbl_graph.data.frame}},\code{\link[tidygraph]{activate}}
}
