% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_terms}
\alias{get_terms}
\title{get_terms}
\usage{
get_terms(
  data,
  groups,
  text,
  algorithm = "RAKE",
  phrase_pattern = "(A|N)*N(P+D*(A|N)*N)*",
  n_cores = 1,
  show_progress = TRUE
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{groups}{PARAM_DESCRIPTION}

\item{text}{PARAM_DESCRIPTION}

\item{algorithm}{PARAM_DESCRIPTION, Default: 'RAKE'}

\item{phrase_pattern}{PARAM_DESCRIPTION, Default: '(A|N)\emph{N(P+D}(A|N)\emph{N)}'}

\item{n_cores}{PARAM_DESCRIPTION, Default: 1}

\item{show_progress}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}}
\code{\link[dplyr]{rename}}
\code{\link[udpipe]{udpipe_load_model}},\code{\link[udpipe]{udpipe_annotate}},\code{\link[udpipe]{udpipe}},\code{\link[udpipe]{keywords_rake}},\code{\link[udpipe]{keywords_collocation}},\code{\link[udpipe]{as_phrasemachine}},\code{\link[udpipe]{keywords_phrases}}
}
