% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfidf.R
\name{get_tfidf}
\alias{get_tfidf}
\title{get_tfidf}
\usage{
get_tfidf(data, groups, keywords, sep = ";", n_keywords = 15)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{groups}{PARAM_DESCRIPTION}

\item{keywords}{PARAM_DESCRIPTION}

\item{sep}{PARAM_DESCRIPTION, Default: ';'}

\item{n_keywords}{PARAM_DESCRIPTION, Default: 15}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()) {
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{rename}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{count}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{top_n}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{mutate-joins}}
\code{\link[tidyr]{separate_rows}}
\code{\link[stringr]{str_trim}}
\code{\link[tidytext]{bind_tf_idf}}
}
