% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topological_positions.R
\name{get_topological_positions}
\alias{get_topological_positions}
\title{get_topological_positions}
\usage{
get_topological_positions(
  data,
  id,
  group,
  PY = PY,
  TI = TI,
  CR = CR,
  min.citations = 1
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION}

\item{PY}{PARAM_DESCRIPTION, Default: PY}

\item{TI}{PARAM_DESCRIPTION, Default: TI}

\item{CR}{PARAM_DESCRIPTION, Default: CR}

\item{min.citations}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{rename}},\code{\link[dplyr]{mutate}},
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},
\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},
\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{bind}},
\code{\link[dplyr]{desc}}
\code{\link[stringr]{str_locate}}
\code{\link[tidyr]{pivot_longer}}
\code{\link[stats]{sd}}
}
