% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2html.R
\name{img2html}
\alias{img2html}
\title{FUNCTION_TITLE}
\usage{
img2html(
  inputPath = "/home/roney/Desktop/gallery/2019_MartinaFlorence/",
  pattern = "*.jpg",
  rows = 1,
  outputPath = "../../gallery/2020_MartinaFlorence/",
  output = "~/output.html"
)
}
\arguments{
\item{inputPath}{PARAM_DESCRIPTION, Default: '/home/roney/Desktop/gallery/2019_MartinaFlorence/'}

\item{pattern}{PARAM_DESCRIPTION, Default: '*.jpg'}

\item{rows}{PARAM_DESCRIPTION, Default: 1}

\item{outputPath}{PARAM_DESCRIPTION, Default: '../../gallery/2020_MartinaFlorence/'}

\item{output}{PARAM_DESCRIPTION, Default: '~/output.html'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
