% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes_plots.R
\name{indexes_plots}
\alias{indexes_plots}
\title{Create CCT or Entropy Visualization Plots}
\usage{
indexes_plots(data, group_name, start_year, end_year, method = "cct")
}
\arguments{
\item{data}{Data from calculate_cct or calculate_entropy function}

\item{group_name}{Specific group to visualize}

\item{start_year}{Starting year for x-axis}

\item{end_year}{Ending year for x-axis}

\item{method}{Character string indicating the method: "cct" or "entropy"}
}
\value{
A plotly object with combined plots
}
\description{
Create CCT or Entropy Visualization Plots
}
\keyword{internal}
