% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_terms.R
\name{load_udpipe_model}
\alias{load_udpipe_model}
\title{Load UDPipe model with on-demand downloading}
\usage{
load_udpipe_model(model_name = "english", model_dir = ".")
}
\arguments{
\item{model_name}{Name of the model to load (default: "english")}

\item{model_dir}{Directory where models are stored (default: current directory)}
}
\value{
A UDPipe model object
}
\description{
Load UDPipe model with on-demand downloading
}
\keyword{internal}
