% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{normalize_column_names}
\alias{normalize_column_names}
\title{Normalize column names across formats}
\usage{
normalize_column_names(data, format)
}
\description{
Normalize column names across formats
}
\keyword{internal}
