% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_formation.R
\name{plot_group_formation_3d}
\alias{plot_group_formation_3d}
\title{Plot a group's formation (3D confluence)}
\usage{
plot_group_formation_3d(formation, ...)
}
\arguments{
\item{formation}{A \code{\link[=sniff_group_formation]{sniff_group_formation()}} object.}

\item{...}{Passed to \code{\link[=plot_trajectory_formation_3d]{plot_trajectory_formation_3d()}}.}
}
\value{
A plotly object.
}
\description{
Thin wrapper over \code{\link[=plot_trajectory_formation_3d]{plot_trajectory_formation_3d()}} for a
\code{\link[=sniff_group_formation]{sniff_group_formation()}} object.
}
\seealso{
\code{\link[=sniff_group_formation]{sniff_group_formation()}}, \code{\link[=plot_trajectory_formation_3d]{plot_trajectory_formation_3d()}}
}
