% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_trajectory_dag.R
\name{print.birddog_dag}
\alias{print.birddog_dag}
\title{Print a soft trajectory DAG}
\usage{
\method{print}{birddog_dag}(x, ...)
}
\arguments{
\item{x}{A \code{birddog_dag} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a soft trajectory DAG
}
\keyword{internal}
