% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_trajectory_braid.R
\name{print.birddog_flow}
\alias{print.birddog_flow}
\title{Print a flow-trajectory object}
\usage{
\method{print}{birddog_flow}(x, ...)
}
\arguments{
\item{x}{A \code{birddog_flow} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a flow-trajectory object
}
\keyword{internal}
