% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{read_lines_multiple}
\alias{read_lines_multiple}
\title{Read lines from single or multiple files}
\usage{
read_lines_multiple(file)
}
\description{
Read lines from single or multiple files
}
\keyword{internal}
