% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{read_wos_plain}
\alias{read_wos_plain}
\title{Read Web of Science plain text files}
\usage{
read_wos_plain(file, normalized_names = TRUE)
}
\arguments{
\item{file}{Character scalar or vector. Path(s) to a WoS export file, or a
single URL (\verb{http://} or \verb{https://}) pointing to a WoS export.}

\item{normalized_names}{Logical. If \code{TRUE} (default), use standardized column
names when possible; if \code{FALSE}, keep original WoS field tags.}
}
\description{
Read Web of Science plain text files
}
\keyword{internal}
