% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_lineage.R
\name{sniff_groups_cumulative_attributes}
\alias{sniff_groups_cumulative_attributes}
\title{Extract attributes from cumulative groups}
\usage{
sniff_groups_cumulative_attributes(
  cummulative_network,
  min_group_size = 10,
  top_n_keywords = 3,
  group_to_track = "component1_g01",
  attributes = "groups"
)
}
\description{
Extract attributes from cumulative groups
}
\keyword{internal}
