% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_tfidf.R
\name{sniff_tfidf}
\alias{sniff_tfidf}
\title{sniff_terms}
\usage{
sniff_tfidf(
  data,
  group,
  term,
  frequency,
  separate_rows = FALSE,
  sep = ";",
  n_terms = 15
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION}

\item{term}{PARAM_DESCRIPTION}

\item{frequency}{PARAM_DESCRIPTION}

\item{separate_rows}{PARAM_DESCRIPTION, Default: FALSE}

\item{sep}{PARAM_DESCRIPTION, Default: ';'}

\item{n_terms}{PARAM_DESCRIPTION, Default: 15}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{rename}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{count}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{top_n}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{mutate-joins}}
\code{\link[tidyr]{separate_rows}}
\code{\link[stringr]{str_trim}}
\code{\link[tidytext]{bind_tf_idf}}
}
