% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{split_wos_records}
\alias{split_wos_records}
\title{Split WOS plain text into individual records}
\usage{
split_wos_records(lines)
}
\description{
Split WOS plain text into individual records
}
\keyword{internal}
