% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birddog-package.R
\docType{package}
\name{birddog-package}
\alias{birddog}
\alias{birddog-package}
\title{birddog: sniffing emergence and trajectories in academic papers and patents}
\description{
Tools to detect emergence and trace technological/scientific trajectories
in papers and patents. It reads OpenAlex and Web of Science data, builds
citation-based networks, identifies groups, and summarizes their dynamics.
}
\section{Links}{

\itemize{
\item Website: \url{https://roneyfraga.com/birddog}
\item GitHub: \url{https://github.com/roneyfraga/birddog}
\item Issues: \url{https://github.com/roneyfraga/birddog/issues}
}
}

\seealso{
\url{https://roneyfraga.com/birddog},
\url{https://github.com/roneyfraga/birddog}
}
\author{
\strong{Maintainer}: Roney Fraga Souza \email{roneyfraga@gmail.com} (\href{https://orcid.org/orcid.org/0000-0001-5750-489X}{ORCID})

Other contributors:
\itemize{
  \item Luis Felipe de Souza Rodrigues \email{lfsouza25@gmail.com} [contributor]
}

}
\keyword{package}
