% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_trajectories.R
\name{extract_docs_for_all_groups}
\alias{extract_docs_for_all_groups}
\title{Extract documents for all groups across all time periods}
\usage{
extract_docs_for_all_groups(groups_cumulative, min_group_size = 10)
}
\arguments{
\item{groups_cumulative}{List of cumulative group data}

\item{min_group_size}{Minimum group size filter}
}
\value{
Data frame with document information for all groups
}
\description{
Extract documents for all groups across all time periods
}
\keyword{internal}
