% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups_trajectories.R
\name{jaccard}
\alias{jaccard}
\title{Calculate Jaccard Similarity Between Two Vectors}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}
}
\value{
Jaccard similarity coefficient (between 0 and 1)
}
\description{
Calculate Jaccard Similarity Between Two Vectors
}
\keyword{internal}
