% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{igraph network}
\format{
A igraph network:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\usage{
net
}
\description{
Rede de artigos científicos com exemplos.
}
\keyword{datasets}
