% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff_groups.R
\name{sniff_groups}
\alias{sniff_groups}
\title{sniff_groups}
\usage{
sniff_groups(
  net,
  min_group_size = 10,
  keep_component = c("component01", "component02"),
  cluster_component = "component01",
  algorithm = "louvain"
)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{min_group_size}{PARAM_DESCRIPTION, Default: 10}

\item{keep_component}{PARAM_DESCRIPTION, Default: c("component01", "component02")}

\item{cluster_component}{PARAM_DESCRIPTION, Default: 'component01'}

\item{algorithm}{PARAM_DESCRIPTION, Default: 'louvain'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{cluster_louvain}},\code{\link[igraph]{cluster_walktrap}},
\code{\link[igraph]{cluster_edge_betweenness}},\code{\link[igraph]{cluster_fast_greedy}},
\code{\link[igraph]{as_data_frame}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{context}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{bind}}
\code{\link[stringr]{str_pad}}
\code{\link[tidygraph]{as_tbl_graph.data.frame}},\code{\link[tidygraph]{activate}}
}
